requirements.txt
alabaster==0.7.12
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
docutils==0.17.1
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
indic-nlp-library==0.81
Jinja2==3.1.2
MarkupSafe==2.1.1
miniful==0.0.6
Morfessor==2.0.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pyFUME==0.2.25
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scipy==1.10.0
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tqdm==4.64.1
urllib3==1.26.13
wikiextractor==3.0.6
zipp==3.11.0
